Feat: Update the format code postal with reliable source (#47)
# Context

After extensive research, I found a reliable reference source on
[dataNova](https://datanova.laposte.fr/), the open data portal of La
Poste Groupe.

I retrieved the latest data for the `CodePostal` format updated on 8
February 2025.

However, no history of data has been found on this source.

# Goal

The purpose of this PR is to update the `CodePostal` format with new
versioned data that is more robust and backed by a trusted source.
Additionally, all versions are now stored in a single dedicated file
`versions.py`.